LPS-162251 no need for this interface inside impl
